+++ cli/main.rs
@@ -1219,11 +1219,6 @@ pub fn main() {
   colors::enable_ansi(); // For Windows 10
 
   let args: Vec<String> = env::args().collect();
-  if let Err(err) = standalone::try_run_standalone_binary(args.clone()) {
-    eprintln!("{}: {}", colors::red_bold("error"), err.to_string());
-    std::process::exit(1);
-  }
-
   let flags = flags::flags_from_vec(args);
   if !flags.v8_flags.is_empty() {
     init_v8_flags(&*flags.v8_flags);
