+++ runtime/lib.rs
@@ -20,6 +20,7 @@ pub mod js;
 pub mod metrics;
 pub mod ops;
 pub mod permissions;
+pub mod standalone;
 pub mod resolve_addr;
 pub mod tokio_util;
 pub mod web_worker;
